// x-pack/solutions/security/plugins/security_solution/common/api/entity_analytics/entity_store/resolution/list_primaries.gen.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License
 * 2.0; you may not use this file except in compliance with the Elastic License
 * 2.0.
 */

/*
 * NOTICE: Do not edit this file manually.
 * This file is automatically generated by the OpenAPI Generator, @kbn/openapi-generator.
 *
 * info:
 *   title: Entity Store - List Primary Entities (FIELDS Architecture)
 *   version: 2023-10-31
 */

import { z } from '@kbn/zod';

import { EntityType, InspectQuery } from '../common.gen';
import { EntityField, Asset } from '../entities/common.gen';

export type PrimaryEntity = z.infer<typeof PrimaryEntity>;
export const PrimaryEntity = z.object({
  '@timestamp': z.string().datetime().optional(),
  entity: EntityField,
  user: z.object({}).catchall(z.unknown()).optional(),
  host: z.object({}).catchall(z.unknown()).optional(),
  service: z.object({}).catchall(z.unknown()).optional(),
  asset: Asset.optional(),
  /**
   * Number of entities resolved to this primary (0 if standalone)
   */
  resolved_count: z.number().int(),
});

export type ListPrimaryEntitiesRequestQuery = z.infer<typeof ListPrimaryEntitiesRequestQuery>;
export const ListPrimaryEntitiesRequestQuery = z.object({
  /**
   * Page number (1-indexed)
   */
  page: z.coerce.number().int().min(1).optional().default(1),
  /**
   * Number of items per page
   */
  per_page: z.coerce.number().int().min(1).max(100).optional().default(10),
  /**
   * Field to sort by
   */
  sort_field: z.string().optional().default('@timestamp'),
  /**
   * Sort order
   */
  sort_order: z.enum(['asc', 'desc']).optional().default('desc'),
  /**
   * JSON-encoded filter query
   */
  filter_query: z.string().optional(),
});
export type ListPrimaryEntitiesRequestQueryInput = z.input<typeof ListPrimaryEntitiesRequestQuery>;

export type ListPrimaryEntitiesRequestParams = z.infer<typeof ListPrimaryEntitiesRequestParams>;
export const ListPrimaryEntitiesRequestParams = z.object({
  entityType: EntityType,
});
export type ListPrimaryEntitiesRequestParamsInput = z.input<
  typeof ListPrimaryEntitiesRequestParams
>;

export type ListPrimaryEntitiesResponse = z.infer<typeof ListPrimaryEntitiesResponse>;
export const ListPrimaryEntitiesResponse = z.object({
  primaries: z.array(PrimaryEntity),
  /**
   * Total number of primary entities
   */
  total: z.number().int(),
  inspect: InspectQuery.optional(),
});